mac-2

+++ server/app/data/controllers.py
@@ -16,6 +16,15 @@ def trade_log():
     trade_log_json = trade_log_schema.dumps(trade_log)
     return jsonify(trade_log_json), 200
 
+@data.route("/api/data/holdings", methods=["GET"])
+@login_required
+def holdings():
+    user_id = current_user.id
+    holdings = Holding.query.filter_by(user_id=user_id).all()
+    holdings_schema = Holding_Schema(many=True)
+    holdings_json = holdings_schema.dumps(holdings)
+    return jsonify(holdings_json), 200
+
 @data.route("/api/data/tradeLogInsert", methods=["POST"])
 @login_required
 def trade_log_insert():
@@ -28,7 +37,11 @@ def trade_log_insert():
     price = request.json.get("price", None)
     commission = request.json.get("commission", None)
 
-    holding_quantity, holding_acb, holding_acb_ps = holding_insert_acb(user_id, transaction, ticker, quantity, price, commission)
+    if is_transaction_valid("Insert", user_id, transaction, ticker, int(quantity)) == False:
+        msg = f"Failed to add trade. Adding would bring your {ticker} share count below zero."
+        return jsonify(msg), 404
+
+    holding_quantity, holding_acb, holding_acb_ps = calculate_acb("Insert", user_id, transaction, ticker, quantity, price, commission)
     new_holding = Holding(user_id=user_id, account=account, ticker=ticker, quantity=holding_quantity, average_cost_basis=holding_acb, average_cost_basis_ps=holding_acb_ps)
     db.session.add(new_holding)
     db.session.commit()
@@ -37,18 +50,48 @@ def trade_log_insert():
     db.session.add(new_trade)
     db.session.commit()
 
-    return jsonify("Successfully added new trade"), 200
+    return jsonify("Successfully logged a new trade."), 200
 
 @data.route("/api/data/tradeLogRemove", methods=["POST"])
 @login_required
 def trade_log_remove():
     user_id = current_user.id
     trade_id = request.json.get("trade_id", None)
+
+    existing_trade = Trade_Log.query.filter_by(id=trade_id, user_id=user_id).one()
+    account = existing_trade.account
+    transaction = existing_trade.transaction
+    ticker = existing_trade.ticker
+    quantity = existing_trade.quantity
+    price = existing_trade.price
+    commission = existing_trade.commission
+
+    if is_transaction_valid("Remove", user_id, transaction, ticker, int(quantity)) == False:
+        msg = f"Failed to remove trade. Adding would bring your {ticker} share count below zero."
+        return jsonify(msg), 404
+
+    holding_quantity, holding_acb, holding_acb_ps = calculate_acb("Remove", user_id, transaction, ticker, quantity, price, commission)
+    new_holding = Holding(user_id=user_id, account=account, ticker=ticker, quantity=holding_quantity, average_cost_basis=holding_acb, average_cost_basis_ps=holding_acb_ps)
+    db.session.add(new_holding)
+    db.session.commit()
+
     Trade_Log.query.filter_by(id=trade_id, user_id=user_id).delete()
     db.session.commit()
-    return jsonify("Successfully deleted trade"), 200
 
-def holding_insert_acb(user_id, transaction, ticker, quantity, price, commission):
+    return jsonify("Successfully removed a trade."), 200
+
+def is_transaction_valid(operation, user_id, transaction, ticker, quantity):
+    if (operation == "Insert" and transaction == "Sell") or (operation == "Remove" and transaction == "Buy"):
+        try:
+            pre_holding = Holding.query.filter_by(user_id=user_id, ticker=ticker).one()
+            pre_quantity = pre_holding.quantity
+            if(pre_quantity < quantity):
+                return False
+        except:
+            return False
+    return True
+    
+def calculate_acb(operation, user_id, transaction, ticker, quantity, price, commission):
     try:
         pre_holding = Holding.query.filter_by(user_id=user_id, ticker=ticker).one()
         pre_quantity = pre_holding.quantity
@@ -65,13 +108,24 @@ def holding_insert_acb(user_id, transaction, ticker, quantity, price, commission
     price = Decimal(price)
     commission = Decimal(commission)
 
-    if transaction == "Buy":
-        post_quantity = pre_quantity + quantity
-        post_acb = pre_acb + (quantity * price) + commission
+    if operation == "Insert":
+        if transaction == "Buy":
+            post_quantity = pre_quantity + quantity
+            post_acb = pre_acb + ((quantity * price) + commission)
+        else:
+            post_quantity = pre_quantity - quantity
+            post_acb = pre_acb - (quantity * pre_acb/pre_quantity)
     else:
-        post_quantity = pre_quantity - quantity
-        post_acb = pre_acb - (quantity * pre_acb/pre_quantity)
+        if transaction == "Buy":
+            post_quantity = pre_quantity - quantity
+            post_acb = pre_acb - ((quantity * price) + commission)
+        else:
+            post_quantity = pre_quantity + quantity
+            post_acb = pre_acb + (quantity * pre_acb/pre_quantity)
 
-    post_acb_ps = post_acb / post_quantity
+    if post_quantity != Decimal(0):
+        post_acb_ps = post_acb / post_quantity
+    else:
+        post_acb_ps = Decimal(0)
 
     return int(post_quantity), Decimal(post_acb), Decimal(post_acb_ps)
